+++ src/tests/image_ops.rs
@@ -128,7 +128,7 @@ fn image_ops() {
         // Write to src:
         core::enqueue_write_image(proque.queue(), &img_src, true, 
             origin, region, 0, 0,
-            &vec, None::<&core::EventList>, None).unwrap();
+            &vec, None, None).unwrap();
 
         // Add from src to dst:
         kernel_add.enq().expect("[FIXME]: HANDLE ME!");
@@ -138,7 +138,7 @@ fn image_ops() {
         // Read into vec:
         unsafe { core::enqueue_read_image(proque.queue(), &img_dst, true, 
             origin, region, 0, 0,
-            &mut vec, None::<&core::EventList>, None).unwrap(); }
+            &mut vec, None, None).unwrap(); }
 
         // Just to make sure read is complete:
         proque.queue().finish();
@@ -154,13 +154,13 @@ fn image_ops() {
         let cur_pixel = [cur_val, cur_val, cur_val, cur_val];
         kernel_fill_src.set_arg_vec_named("pixel", &cur_pixel).unwrap().enq().expect("[FIXME]: HANDLE ME!");
 
-        core::enqueue_copy_image::<i32, _>(proque.queue(), &img_src, &img_dst, 
-            origin, origin, region, None::<&core::EventList>, None).unwrap();
+        core::enqueue_copy_image::<i32>(proque.queue(), &img_src, &img_dst, 
+            origin, origin, region, None, None).unwrap();
 
         // Read into vec:
         unsafe { core::enqueue_read_image(proque.queue(), &img_dst, true, 
             origin, region, 0, 0,
-            &mut vec, None::<&core::EventList>, None).unwrap(); }
+            &mut vec, None, None).unwrap(); }
 
         // Just to make sure read is complete:
         proque.queue().finish();